Added global config file
In order to change to another minimum, it is interesting to change
penalty points. The final solution for the penalty points problem
may be it is a linear combination of the normalized vectors.

In the same way solutions for each of the normalized penalt points
vectors can be taken as "models" to recombine by the mutation
controller.

Mutation controller walks toward analyzing the
effect of single mutations and (maybe) multiple combinations.

All of the previous ideas need that the config constant become
variables. And that's what this commit do.